Multiple hardware accelerators for layer update, spike generation done in software

+++ libs/sbs_neural_network/inc/mt19937int.h
@@ -1,3 +1,3 @@
 
-void sgenrand(unsigned long seed);
-unsigned long genrand();
+void sgenrand(unsigned int seed);
+unsigned int genrand();
